requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
flask-limiter==3.5.0
typing-extensions==4.8.0
